arreglado la carpeta de otro proyecto y mostrar solo a los entrenadores en operadores
