--- proyecto_gym-desarrollo/manage.py
@@ -1,6 +1,0 @@
-from app import inicializador_app
-
-app = inicializador_app()
-
-if __name__ == '__main__':
-    app.run(debug=True)
