--- proyecto_gym-desarrollo/static/css/login.css
@@ -1,170 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Oswald:wght@400;500&display=swap');
-
-* {
-    box-sizing: border-box;
-    transition: all 0.3s;
-    font-family: 'Oswald', sans-serif;
-    text-transform: uppercase;
-}
-:root{
-    --gris_claro:#f1f3f9;
-    --azul_claro:#3f4464;
-    --azul_oscuro:#1f2232;
-    --amarillo:#ffc857;
-}
-body, html {
-    margin: 0;
-    padding: 0;
-}
-
-main#inicio {
-    background-color: var(--gris_claro);
-    height: 100vh;
-    width: 100vw;
-    position: relative;
-    display: flex;
-   align-items: center;
-   justify-content: center;
- 
-}
-
-img {
-    width: 100px;
-}
-
-.contenedor {
-    background-color: var(--azul_oscuro);
-    position: absolute;
-    display: inline-block;
-    align-items: center;
-    justify-content: space-around;
-    padding: 20px;
-    width: 400px;
-    height: 400px;
-    border-radius: 10px;
-   
-    z-index: 1;
-    
-
-}
-
-.titulo-login {
-    text-align: center;
-    color: var(--gris_claro);
-    font-weight: 500;
-    margin: 80px 0 0 0;
-}
-
-.formulario {
-    /* background-color: #3F4464; */
-    background-color:var(--azul_oscuro);
-    border-radius: 6px;
-    border: none;
-    padding: 40px;
-    max-width: 340px;
-    min-height: 45vh;
-    min-width: 300px;
-    width: 100%;
-    position: relative;
-}
-
-.input_contenedor input {
-
-    display: block;
-    width: 100%;
-    padding: 10px;
-    border: none;
-    /* background-color: #f1f3f9; */
-    background-color: var(--azul_claro);
-    border-radius: 2px;
-}
-
-#escribir::placeholder {
-    color: #f1f3f9cb;
-}
-
-#contrasena::placeholder {
-    /* color: #f1f3f9; */
-    color: #f1f3f9cb;
-}
-
-.input_contenedor {
-    margin-top: 10%;
-}
-
-input {
-    margin-bottom: 30px;
-}
-
-.input-container {
-    position: relative;
-  }
-  
-  .input-container i {
-    position:absolute;
-    top: 50%;
-    left: 90%;
-    transform: translateY(-50%);
-    /* color: #3F4464; */
-    color: var(--gris_claro);
-  }
-  
-.logo-gym {
-    position: absolute;
-    top: -80px;
-    left: 50%;
-    background-color: var(--amarillo);
-    border-radius: 50%; 
-    padding: 20px;
-    transform: translateX(-50%);
-    margin: 0;
-    
-    
-}
-
-.boton {
-    display: flex;
-    justify-content: center;
-}
-
-.boton_entrar {
-    background-color: var(--amarillo);
-    border: none;
-    border-radius: 2px;
-    width: 100%;
-    padding: 8px;
-    cursor: pointer;
-    letter-spacing: 2px;
-    font-weight: 500;
-    color: #473819;
-}
-
-
-
-.btssesion{
-    background-color: var(--amarillo);
-    border: none;
-    border-radius: 0px 0px 5px 5px;
-    padding: 5px;
-    width: 140px;
-    cursor: pointer;
-    color: #473819;
-    
-    
-}
-.tipo{
-    
-    bottom: -30px;
-    position: absolute;
-    width: 100%;
-    justify-content: space-around;
-    display: flex;
-    justify-content: flex-end;
-    right: 0;
-    padding: 0 17px;
-    z-index: 3;
-}
-
-
-
-
