--- proyecto_gym-desarrollo/static/css/navbar.css
@@ -1,252 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@500&display=swap');
-
-* {
-    box-sizing: border-box;
-    transition: all 0.3s ease-in-out;
-}
-
-body, html {
-    margin: 0;
-    padding: 0;
-}
-
-body {
-    font-size: 16px;
-    font-family: 'Poppins', sans-serif;
-}
-
-.contenedor {
-    background-color: #D6D6D6;
-    min-height: 100vh;
-}
-
-.navbar {
-    background-color: #333533;
-    color: #fff;
-    display: flex;
-    justify-content:flex-start;
-    align-items: center;
-    padding: 10px 20px;
-    height: 90px;
-}
-
-
-.logo a {
-    text-decoration: none;
-    color: #fff;
-    font-weight: bold;
-    font-size: 24px;
-}
-
-.menu {
-    list-style: none;
-    display: flex;
-}
-
-.menu li {
-    margin-right: 30px;
-    margin-top: 30px;
-}
-
-.links {
-    margin-top: 100px;
-}
-
-.navbar .btn {
-    margin-left: auto;
-}
-
-.btn a {
-    text-decoration: none;
-    color: #000;
-    margin: 0 50px 0 30px;
-    padding: 10px;
-    background-color: #FFD100;
-    border-radius: 6px;
-    box-shadow: 0px 4px 11px 0px rgba(0, 0, 0, 0.25);
-}
-
-.btn a i {
-    margin-left: 5px; /* Espacio entre el icono y el texto */
-}
-
-.btn a:hover {
-    background-color: #d3ae0b;
-    box-shadow: 0 0 15px rgba(255, 208, 0, 0.288);
-}
-
-.menu a {
-    text-decoration: none;
-    color: #fff;
-    font-size: 18px;
-}
-
-.menu-toggle {
-    display: none;
-    cursor: pointer;
-    margin: 0 50px 0 10px;
-}
-
-.bar {
-    width: 30px;
-    height: 4px;
-    background-color: #fff;
-    margin: 3px 0;
-    transition: 0.4s;
-}
-
-
-    .menu-toggle {
-        display: block;
-    }
-
-    .menu {
-        display: none;
-        position: absolute;
-        top: 0px;
-        left: 0px;
-        background-color: #333533;
-        flex-direction: column;
-        align-items: flex-start;
-        text-align: start;
-        padding: 0 0 0 30px;
-    }
-
-    .menu.active {
-        display: flex;
-        height: 100vh;
-        width: 300px;
-        margin: 0;
-    }
-
-    .menu-toggle.active .bar:nth-child(1) {
-        transform: translateY(5px) rotate(45deg);
-    }
-
-    .menu-toggle.active .bar:nth-child(2) {
-        opacity: 0;
-    }
-
-    .menu-toggle.active .bar:nth-child(3) {
-        transform: translateY(-9px) rotate(-45deg);
-    }
-
-
-.salir {
-    position: absolute;
-    top: 94%;
-    
-}
-    
-.btn_cerrar {
-    background-color: none;
-    color: #000;
-    padding: 10px;
-    background-color: #d3ae0b;
-    border-radius: 6px;
-    box-shadow: 0px 4px 11px 0px rgba(0, 0, 0, 0.25);
-}
-    
-.btn_cerrar:hover {
-    background-color: #000;
-}
-
-/* ---------------- MODAL USUARIO REGISTRO -----------------*/
-
-.modal_user {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: #111111bd;
-    display: flex;
-    opacity: 0;
-    pointer-events: none;
-    transition: opacity .6s;
-    --transform: translateY(-100vh);
-    --transition: transform .8s;
-    }
-    
-    .modal_info {
-        position: fixed;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-        background-color: #111111bd;
-        display: flex;
-        opacity: 0;
-        pointer-events: none;
-        transition: opacity .6s;
-        --transform: translateY(-100vh);
-        --transition: transform .8s;
-        }
-    
-    .modal--show {
-    opacity: 1;
-    pointer-events: unset;
-    transition: opacity .6s;
-    --transform: translateY(0);
-    }
-    
-    .modal_container {
-    overflow-y: scroll;
-    width: 90%;
-    max-width: 400px;
-    max-height: 90%;
-    border-radius: 6px;
-    padding: 3em 2.5em;
-    background-color: #fff;
-    margin: auto;
-    display: grid;
-    gap: 1em;
-    place-items: center;
-    grid-auto-columns: 100%;
-    transform: var(--transform);
-    transition: var(--transition);
-    }
-    
-    .modal_titulo {
-    font-size: 2.5em;
-    }
-    
-    .modal_texto {
-    margin-bottom: 10px;
-    }
-    
-    
-    .modal_close_user {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    text-decoration: none;
-    color: #000;
-    padding: 5px 10px;
-    border-radius: 8px;
-    transition: background-color 0.3s;
-    }
-    
-    .modal_close_user:hover {
-    color: #fff;
-    border-radius: 8px;
-    background-color: #5a5a5a;
-    }
-
-/* ------------- RESPONSIVE ------------------- */
-
-/* @media screen and (max-width: 768px) {
-    .navbar {
-        position: relative;
-    }
-
-    .btn {
-        display: flex;
-        position: absolute;
-        top: 100px;
-        gap: 10px;
-    }
-    .btn a {
-        margin: 0;
-    }
-} */
