--- proyecto_gym-desarrollo/templates/dashboard/cabecera.html
@@ -1,71 +1,0 @@
-<body>
-    <div class="contenedor">
-    <nav class="navbar">
-        <ul class="menu">
-            <div class="links">
-                <li><a href="/inicio">Inicio</a></li>
-                {% if session["rol"] == 'super_admin' or session["rol"] == 'administrador' %}
-                <li><a href="/operadores">Operadores</a></li>
-                {% endif %}
-                <li><a href="/inventario">Inventario de Productos</a></li>
-                <li><a href="/afiliados">Afiliados</a></li>
-                <li><a href="/membresias">Membresias</a></li>
-                <li><a href="#">Servicios</a></li>
-                <li><a href="#">Contacto</a></li>
-                </div>
-                <div class="salir">
-                    <a class="btn_cerrar" href="/dashboard/cerrar">Cerrar Sesion</a>
-                </div>
-            </ul>
-        
-        <div class="menu-toggle" id="menu-toggle">
-            <div class="bar"></div>
-            <div class="bar"></div>
-            <div class="bar"></div>
-        </div>
-        <div class="logo">
-            <a href="/inicio">CONTROL GYM</a>
-        </div>
-        <div class="btn">
-            <a href="/inventario">Venta producto <i class="fa-solid fa-cart-shopping"></i> </a>
-            <a href="/afiliados" class="boton_user">Registrar usuario <i class="fa-solid fa-user-plus"></i> </a>
-        </div>
-    </nav>
-
-<!--     <section class="modal_user">
-        <div class="modal_container">
-            <a href="#" class="modal_close_user"><i class="fa-solid fa-xmark fa-xl"></i></a>
-            <div class="container_form">
-                <div class="form_container">
-                    <h2 class="titulo">AGREGAR NUEVO USUARIO</h2>
-                    <form id="formulario" action="/afiliados/agregarAfiliado" method="POST">
-                        <div class="content">
-                            
-                            <input type="text" class="inputText" id="cedula" name="cedula" placeholder="Cedula" required> 
-                            
-                            <input type="text" class="inputText" id="nombres" name="nombres" placeholder="Nombres" required>
-        
-                            <input type="text" class="inputText" id="apellidos" name="apellidos" placeholder="Apellidos" required>
-                            
-                            <label for="fecha_nac" style="font-size: 14px;">Fecha de Nacimiento:</label>
-                            <input type="text" class="inputText" id="fecha_nac" name="fecha_nac" placeholder="año/mes/dia" required>
-        
-                            <input type="tel" class="inputText" id="telefono" name="telefono" placeholder="Telefono Movil" required>
-        
-                            <input type="email" class="inputText" id="email" name="email" placeholder="Correo Electronico" required>
-        
-                            <input type="number" class="inputText" id="nfc" name="nfc" placeholder="tarjeta nfc" required>
-                            
-                            <label for="membresia" style="font-size: 14px;">Membresia:</label>
-                            <select class="inputText" name="membresia" id="membresia" required>
-                                {% for resul in resulMem %}
-                                <option value="{{ resul[0] }}">{{ resul[1] }}</option>
-                                {% endfor %}
-                            </select>                                 
-                        </div>
-                        <button type="submit" class="buttonSumit">Agregar</button>
-                    </form>
-                </div>
-            </div>
-        </div>
-        </section> -->
