--- proyecto_gym-desarrollo/templates/dashboard/edit_usuario.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/navbar.css') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/infOperadores.css') }}">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.0/css/all.min.css">
-    <title>Control Gym</title>
-</head>
-
-{% include 'dashboard/cabecera.html' %}
-
-<div class="titulo_operadores">
-    <h2>DATOS DE LA MEMBRESIA</h2>
-</div>
-
-<div class="contenedorForm">
-    <form id="formulario" action="/membresias/editarMembresia" method="POST">
-        <div class="content">
-           
-            <label for="nombre">ID: {{editMembresia[0]}}</label>
-            <input type="text" class="inputText" value="{{editMembresia[0]}}" id="txtID" name="txtID" hidden>
-            <br>
-            <br>
-            <label for="nombre">Nombre:</label>
-            <input type="text" class="inputText" value="{{editMembresia[1]}}" pattern="[a-zA-Z]{2,20}" minlength="2" maxlength="20" id="nombre" name="nombre" required>
-
-            <label for="nombre">Duracion:</label>
-            <input type="text" class="inputText" value="{{editMembresia[2]}}" pattern="[0-9]{1,3}" minlength="1" maxlength="3" id="duracion" name="duracion" required>
-
-            <label for="nombre">Precio:</label>
-            <input type="number" class="inputText" value="{{editMembresia[3]}}" pattern="[0-9]" minlength="1" id="precio" name="precio" required>
-            
-        </div>
-
-            <div class="btn_accion">
-                <button type="submit" class="buttonEdit">Editar</button>
-                <!-- <a href="" class="buttonEdit">Editar</a> -->
-                <a href="/membresias" class="btn_desactivar">Cancelar</a>
-            </div>
-
-    </form>
-    
-</div>
-
-{% include 'dashboard/pie.html' %}
-
