--- proyecto_gym-desarrollo/templates/dashboard/inventario_productos.html
@@ -1,115 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/navbar.css') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/operadores.css') }}">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.0/css/all.min.css">
-    <title>Control Gym</title>
-</head>
-
-{% include 'dashboard/cabecera.html' %}
-
-<div class="titulo_operadores">
-    <h2>INVENTARIO</h2>
-</div>
-
-{% if not session["rol"] == 'entrenador' %}
-<div class="boton_agregar_pro">
-    <a href="/categorias" class="boton_cat">Agregar Categoria</a>
-    <a href="" class="boton_ope">Agregar Producto <i class="fa-solid fa-plus"></i></a>
-</div>
-{% endif %}
-
-<div class="container_table">
-<table>
-    <thead>
-        <tr>
-            <!-- <th>Usuario</th> -->
-            <th>Nombre</th>
-            <th>Categoria</th>
-            <!-- <th>Cedula</th> -->
-            <!-- <th>Correo</th> -->
-            <th>Precio de Compra</th>
-            <th>Precio de Venta</th>
-            <th>Cantidad</th>
-            <!-- <th>User Registro</th> -->
-            {% if not session["rol"] == 'entrenador' %}
-            <th>Acciones</th>
-            {% endif %}
-        </tr>
-    </thead>
-    <tbody>
-        {% for resul in productos %}
-        <tr>
-            <td>{{resul[1]}}</td>
-            <td>{{resul[2]}}</td>
-            <td>{{resul[3]}}</td>
-            <td>{{resul[4]}}</td>
-            <td>{{resul[5]}}</td>
-            {% if not session["rol"] == 'entrenador' %}
-            <td class="acciones">
-                
-            <a onclick="return confirm('Seguro quiere eliminar este producto?')" class="btn_eliminar" href="/inventario/desactivar/{{resul[0]}}"><i class="fa-solid fa-trash"></i></a> 
-
-            <a class="btn_edit" href="/inventario/infoEditProducto/{{resul[0]}}"><i class="fa-solid fa-pen-to-square" style="color: #ffffff;"></i></i></a> 
-            </td>
-            {% endif %}
-        </tr>
-        {% endfor %}
-
-    </tbody>
-</table>
-</div>
-
-
-<!-- {% if mensaje %}
-<section class="modal">
-    <div class="modal_container">
-        <a href="#" class="modal_close"><i class="fa-solid fa-xmark fa-xl"></i></a>
-        <div class="container_form">
-            <div class="form_container">
-                <h2 class="titulo">{{ mensaje }}</h2>
-            </div>
-        </div>
-    </div>
-</section>
-{% endif %} -->
-
-
-<section class="modal">
-<div class="modal_container">
-    <a href="#" class="modal_close"><i class="fa-solid fa-xmark fa-xl"></i></a>
-    <div class="container_form">
-        <div class="form_container">
-            <h2 class="titulo">AGREGAR UN PRODUCTO</h2>
-            <form id="formulario" action="/inventario/agregarProducto" method="POST">
-                <div class="content">
-                    
-                    <input type="text" class="inputText" id="nombre" name="nombre" pattern="[A-Za-z0-9\s]{4,30}" minlength="4" maxlength="30" placeholder="Nombre" required>
-                    
-                    <select class="inputText" name="categoria" id="categoria" required>
-                        {% for cate in categorias %}
-                        <option value="{{cate[0]}}">{{cate[1]}}</option>
-                        {% endfor %}
-                    </select>
-
-                    <input type="number" class="inputText" id="precio_compra" name="precio_compra" pattern="[0-9]{2,8}" minlength="2" maxlength="8" placeholder="Precio de compra" required>
-                    
-                    <input type="number" class="inputText" id="precio_venta" name="precio_venta" pattern="[0-9]{2,8}" minlength="2" maxlength="8" placeholder="Precio de Venta" required>              
-                    
-                    <input type="number" class="inputText" id="cantidad" name="cantidad"pattern="[0-9]{2,6}" minlength="2" maxlength="6" placeholder="Cantidad" required>
-                    
-                    <!-- <label for="imagen" style="font-size: 14px;">Imagen del producto:</label>
-                    <input type="file" class="inputText" id="imagen" name="imagen" placeholder=""> -->
-                </div>
-                <button type="submit" class="buttonSumit">Agregar</button>
-            </form>
-        </div>
-    </div>
-</div>
-</section>
-
-
-{% include 'dashboard/pie.html' %}
