--- proyecto_gym-desarrollo/templates/dashboard/membresias.html
@@ -1,98 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/navbar.css') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/operadores.css') }}">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.0/css/all.min.css">
-    <title>Control Gym</title>
-</head>
-
-{% include 'dashboard/cabecera.html' %}
-
-<div class="titulo_operadores">
-    <h2>MEMBRESIAS</h2>
-</div>
-
-{% if not session["rol"] == 'entrenador' %}
-<div class="boton_registrar_ope">
-    <a href="" class="boton_ope">Agregar Membresia<i class="fa-solid fa-plus"></i></a>
-</div>
-{% endif %}
-
-<div class="container_table">
-<table>
-    <thead>
-        <tr>
-            <!-- <th>Usuario</th> -->
-            <th>ID</th>
-            <!-- <th>Cedula</th> -->
-            <th>Nombre</th>
-            <th>Duracion</th>
-            <th>Precio</th>
-            {% if not session["rol"] == 'entrenador' %}
-            <th>Acciones</th>
-            {% endif %}
-        </tr>
-    </thead>
-    <tbody>
-        {% for resul in membresias %}
-        <tr>
-            <td>{{ resul[0] }}</td>
-            <td>{{ resul[1] }}</td>
-            <td>{{ resul[2] }} Dias</td>
-            <td>$ {{ resul[3] }}</td>
-            {% if not session["rol"] == 'entrenador' %}
-            <td class="acciones">
-                
-            <a onclick="return confirm('Seguro quiere eliminar esta membresia?')" class="btn_eliminar" href="/membresias/desactivarMembresia/{{resul[0]}}"><i class="fa-solid fa-trash"></i></a> 
-
-            <a class="btn_edit" href="/membresias/infoEdit/{{resul[0]}}"><i class="fa-solid fa-pen-to-square" style="color: #ffffff;"></i></a> 
-
-            </td>
-            {% endif %}
-        </tr>
-        {% endfor %}
-    </tbody>
-</table>
-</div>
-
-
-<!-- {% if mensaje %}
-<section class="modal">
-    <div class="modal_container">
-        <a href="#" class="modal_close"><i class="fa-solid fa-xmark fa-xl"></i></a>
-        <div class="container_form">
-            <div class="form_container">
-                <h2 class="titulo">{{ mensaje }}</h2>
-            </div>
-        </div>
-    </div>
-</section>
-{% endif %} -->
-
-
-<section class="modal">
-<div class="modal_container">
-    <a href="#" class="modal_close"><i class="fa-solid fa-xmark fa-xl"></i></a>
-    <div class="container_form">
-        <div class="form_container">
-            <h2 class="titulo">AGREGAR UNA MEMBRESIA</h2>
-            <form id="formulario" action="/membresias/agregarMembresia" method="POST">
-                <div class="content">
-                    
-                    <input type="text" class="inputText" id="nombre" name="nombre" pattern="[a-z]{2,20}" minlength="2" maxlength="20" placeholder="Ingresa el nombre" required>
-                    <input type="text" class="inputText" id="duracion" name="duracion" pattern="[0-9]{1,3}" minlength="1" maxlength="3" placeholder="Ingresa la duracion en dias" required>
-                    <input type="number" class="inputText" id="precio" name="precio" pattern="[0-9]" minlength="1" placeholder="Ingresa el precio" required>
-                    
-                </div>
-                <button type="submit" class="buttonSumit">Agregar</button>
-            </form>
-        </div>
-    </div>
-</div>
-</section>
-
-
-{% include 'dashboard/pie.html' %}
