--- proyecto_gym-desarrollo/templates/dashboard/operadores.html
@@ -1,114 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/navbar.css') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/operadores.css') }}">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.0/css/all.min.css">
-    <title>Control Gym</title>
-</head>
-
-{% include 'dashboard/cabecera.html' %}
-
-<div class="titulo_operadores">
-    <h2>LISTA DE OPERADORES</h2>
-</div>
-
-<div class="boton_registrar_ope">
-    <a href="" class="boton_ope">Agregar Operador <i class="fa-solid fa-plus"></i></a>
-</div>
-
-<div class="container_table">
-<table>
-    <thead>
-        <tr>
-            <!-- <th>Usuario</th> -->
-            <th>Nombres</th>
-            <th>Apellidos</th>
-            <!-- <th>Cedula</th> -->
-            <th>Telefono</th>
-            <!-- <th>Correo</th> -->
-            <th>Cargo</th>
-            <!-- <th>Fecha de Registro</th>
-            <th>User Registro</th> -->
-            <th>Mas Info</th>
-        </tr>
-    </thead>
-    <tbody>
-        {% for resul in operadores %}
-        <tr>
-            <!-- <td>{{ resul[0] }}</td> -->
-            <td>{{ resul[1] }}</td>
-            <td>{{ resul[2] }}</td>
-            <!-- <td>{{ resul[3] }}</td> -->
-            <td>{{ resul[4] }}</td>
-            <!-- <td>{{ resul[5] }}</td> -->
-            <td>{{ resul[6] }}</td>
-            <!-- <td>{{ resul[7] }}</td>
-            <td>{{ resul[8] }}</td> -->
-            <td class="acciones">{% if session["rol"] == 'super_admin' %} <a onclick="return confirm('Seguro quiere eliminar este operador?')" class="btn_eliminar" href="/operadores/eliminar/{{resul[0]}}"><i class="fa-solid fa-trash"></i></a> {% endif %} 
-
-            <a class="btn_info" href="/operadores/info/{{resul[0]}}"><i class="fa-solid fa-circle-info" style="color: #ffd100;"></i></a> 
-
-            </td>
-
-        </tr>
-        {% endfor %}
-    </tbody>
-</table>
-</div>
-
-
-<!-- {% if mensaje %}
-<section class="modal">
-    <div class="modal_container">
-        <a href="#" class="modal_close"><i class="fa-solid fa-xmark fa-xl"></i></a>
-        <div class="container_form">
-            <div class="form_container">
-                <h2 class="titulo">{{ mensaje }}</h2>
-            </div>
-        </div>
-    </div>
-</section>
-{% endif %} -->
-
-
-<section class="modal">
-<div class="modal_container">
-    <a href="#" class="modal_close"><i class="fa-solid fa-xmark fa-xl"></i></a>
-    <div class="container_form">
-        <div class="form_container">
-            <h2 class="titulo">REGISTRAR UN OPERADOR</h2>
-            <form id="formulario" action="/operadores/agregarOperador" method="POST">
-                <div class="content">
-                    
-                    <input type="text" class="inputText" id="nombre" name="nombre" pattern="[A-Za-z\s]{2,20}" minlength="2" maxlength="20" placeholder="Ingresa los nombres" required>
-                    
-                    <input type="text" class="inputText" id="apellido" name="apellido" pattern="[A-Za-z\s]{2,20}" minlength="2" maxlength="20" placeholder="Ingresa los apellidos" required>
-                    
-                    <input type="text" class="inputText" id="cedula" name="cedula" pattern="[0-9]{6,15}" minlength="6" maxlength="15" placeholder="Ingresa la cedula" required>
-                    
-                    <input type="email" class="inputText" id="correo" name="correo" pattern="[A-Za-z-_0-9]+@[A-Za-z]\.[a-z]{2,40}" minlength="2" maxlength="40" placeholder="Ingresa el correo" required>
-                    
-                    <input type="tel" class="inputText" id="celular" name="celular" pattern="[0-9]{10}" minlength="10" maxlength="10" placeholder="Ingresa el telefono" required>
-                    
-                    <select class="inputText" name="rol" id="rol" required>
-                        <option value="">Elije un rol</option>
-                        <option value="administrador">Administrador</option>
-                        <option value="entrenador">Entrenador</option>
-                    </select>
-                    
-                    <input type="text" class="inputText" id="usuario" name="usuario" pattern="[A-Za-z-_0-9]{4,30}" minlength="4" maxlength="30" placeholder="Ingresa el usuario" required>
-                    
-                    <input type="password" class="inputText" id="password" name="password" pattern="[A-Za-z-_0-9]{8,20}" minlength="8" maxlength="20" placeholder="Ingresa la contraseña" required>
-                </div>
-                <button type="submit" class="buttonSumit">Registrar</button>
-            </form>
-        </div>
-    </div>
-</div>
-</section>
-
-
-{% include 'dashboard/pie.html' %}
