--- proyecto_gym-desarrollo/templates/dashboard/pie.html
@@ -1,4 +1,0 @@
-</div>
-<script src="{{ url_for('static', filename='js/main.js') }}"></script>
-</body>
-</html>
