+++ static/css/infOperadores.css
@@ -86,6 +86,46 @@ input {
         cursor: pointer;
     }
 
+.buttonSumit {
+    text-decoration: none;
+    padding: 10px 30px;
+    border-radius: 6px;
+    border: 0;
+    font-size: 16px;
+    box-shadow: 1px 1px 10px rgba(0, 0, 0, 0.2);
+    background-color: red;
+    color: #fff;
+    font-family: 'Poppins', sans-serif;
+}
+    
+    .buttonSumit:hover {
+        background-color: rgb(211, 0, 0);
+        transform: scale(1.1);
+        box-shadow: 0 0 15px rgba(212, 0, 0, 0.466);
+        cursor: pointer;
+    }
+
+
+.buttonSumit2 {
+    text-decoration: none;
+    padding: 10px 30px;
+    border-radius: 6px;
+    border: 0;
+    font-size: 16px;
+    box-shadow: 1px 1px 10px rgba(0, 0, 0, 0.2);
+    background-color: #FFD100;
+    color: #fff;
+    font-family: 'Poppins', sans-serif;
+}
+    
+    .buttonSumit2:hover {
+        background-color: #d3ae0b;
+        transform: scale(1.1);
+        box-shadow: 0 0 15px rgba(255, 208, 0, 0.288);
+        cursor: pointer;
+    }
+
+
 .buttonEdit {
     text-decoration: none;
     padding: 10px 30px;
