+++ static/css/navbar.css
@@ -139,15 +139,16 @@ body {
 }
     
 .btn_cerrar {
-    
+    background-color: none;
+    color: #000;
     padding: 10px;
-    background-color: #9D9D9D;
+    background-color: #d3ae0b;
     border-radius: 6px;
     box-shadow: 0px 4px 11px 0px rgba(0, 0, 0, 0.25);
 }
     
 .btn_cerrar:hover {
-    background-color: #696969;
+    background-color: #000;
 }
 
 /* ---------------- MODAL USUARIO REGISTRO -----------------*/
