+++ templates/dashboard/cabecera.html
@@ -27,7 +27,7 @@
             <a href="/inicio">CONTROL GYM</a>
         </div>
         <div class="btn">
-            <a href="">Venta producto <i class="fa-solid fa-cart-shopping"></i> </a>
+            <a href="/inventario">Venta producto <i class="fa-solid fa-cart-shopping"></i> </a>
             <a href="/afiliados" class="boton_user">Registrar usuario <i class="fa-solid fa-user-plus"></i> </a>
         </div>
     </nav>
