+++ templates/dashboard/categoria_productos.html
@@ -37,7 +37,7 @@
             <td>{{resul[1]}}</td>
             <td class="acciones">
                 
-            <a onclick="return confirm('Seguro quiere eliminar este operador?')" class="btn_eliminar" href="/categorias/desactivarCategoria/{{resul[0]}}"><i class="fa-solid fa-trash"></i></a> 
+            <a onclick="return confirm('Seguro quiere eliminar esta categoria?')" class="btn_eliminar" href="/categorias/desactivarCategoria/{{resul[0]}}"><i class="fa-solid fa-trash"></i></a> 
 
             </td>
         </tr>
