+++ templates/dashboard/edit_membresia.html
@@ -24,13 +24,13 @@
             <br>
             <br>
             <label for="nombre">Nombre:</label>
-            <input type="text" class="inputText" value="{{editMembresia[1]}}" pattern="[a-z]{2,20}" minlength="2" maxlength="20" id="nombre" name="nombre">
+            <input type="text" class="inputText" value="{{editMembresia[1]}}" pattern="[a-zA-Z0-9]{2,20}" minlength="2" maxlength="20" id="nombre" name="nombre" required>
 
             <label for="nombre">Duracion:</label>
-            <input type="text" class="inputText" value="{{editMembresia[2]}}" pattern="[0-9]{1,3}" minlength="1" maxlength="3" id="duracion" name="duracion">
+            <input type="text" class="inputText" value="{{editMembresia[2]}}" pattern="[0-9]{1,3}" minlength="1" maxlength="3" id="duracion" name="duracion" required>
 
             <label for="nombre">Precio:</label>
-            <input type="text" class="inputText" value="{{editMembresia[3]}}" pattern="[0-9]" id="precio" name="precio">
+            <input type="number" class="inputText" value="{{editMembresia[3]}}" pattern="[0-9]" minlength="1" id="precio" name="precio" required>
             
         </div>
