+++ templates/dashboard/edit_usuario.html
@@ -24,7 +24,7 @@
             <br>
             <br>
             <label for="nombre">Nombre:</label>
-            <input type="text" class="inputText" value="{{editMembresia[1]}}" pattern="[a-zA-Z0-9]{2,20}" minlength="2" maxlength="20" id="nombre" name="nombre" required>
+            <input type="text" class="inputText" value="{{editMembresia[1]}}" pattern="[a-zA-Z]{2,20}" minlength="2" maxlength="20" id="nombre" name="nombre" required>
 
             <label for="nombre">Duracion:</label>
             <input type="text" class="inputText" value="{{editMembresia[2]}}" pattern="[0-9]{1,3}" minlength="1" maxlength="3" id="duracion" name="duracion" required>
