+++ templates/dashboard/inventario_productos.html
@@ -51,7 +51,7 @@
             {% if not session["rol"] == 'entrenador' %}
             <td class="acciones">
                 
-            <a onclick="return confirm('Seguro quiere eliminar este operador?')" class="btn_eliminar" href="/inventario/desactivar/{{resul[0]}}"><i class="fa-solid fa-trash"></i></a> 
+            <a onclick="return confirm('Seguro quiere eliminar este producto?')" class="btn_eliminar" href="/inventario/desactivar/{{resul[0]}}"><i class="fa-solid fa-trash"></i></a> 
 
             <a class="btn_edit" href="/inventario/infoEditProducto/{{resul[0]}}"><i class="fa-solid fa-pen-to-square" style="color: #ffffff;"></i></i></a> 
             </td>
@@ -87,7 +87,7 @@
             <form id="formulario" action="/inventario/agregarProducto" method="POST">
                 <div class="content">
                     
-                    <input type="text" class="inputText" id="nombre" name="nombre" pattern="[A-Za-z\s]{4,30}" minlength="4" maxlength="30" placeholder="Nombre" required>
+                    <input type="text" class="inputText" id="nombre" name="nombre" pattern="[A-Za-z0-9\s]{4,30}" minlength="4" maxlength="30" placeholder="Nombre" required>
                     
                     <select class="inputText" name="categoria" id="categoria" required>
                         {% for cate in categorias %}
@@ -95,11 +95,11 @@
                         {% endfor %}
                     </select>
 
-                    <input type="text" class="inputText" id="precio_compra" name="precio_compra" pattern="[0-9]{2,8}" minlength="2" maxlength="8" placeholder="Precio de compra" required>
+                    <input type="number" class="inputText" id="precio_compra" name="precio_compra" pattern="[0-9]{2,8}" minlength="2" maxlength="8" placeholder="Precio de compra" required>
                     
-                    <input type="text" class="inputText" id="precio_venta" name="precio_venta" pattern="[0-9]{2,8}" minlength="2" maxlength="8" placeholder="Precio de Venta" required>              
+                    <input type="number" class="inputText" id="precio_venta" name="precio_venta" pattern="[0-9]{2,8}" minlength="2" maxlength="8" placeholder="Precio de Venta" required>              
                     
-                    <input type="text" class="inputText" id="cantidad" name="cantidad"pattern="[0-9]{2,6}" minlength="2" maxlength="6" placeholder="Cantidad" required>
+                    <input type="number" class="inputText" id="cantidad" name="cantidad"pattern="[0-9]{2,6}" minlength="2" maxlength="6" placeholder="Cantidad" required>
                     
                     <!-- <label for="imagen" style="font-size: 14px;">Imagen del producto:</label>
                     <input type="file" class="inputText" id="imagen" name="imagen" placeholder=""> -->
