+++ templates/dashboard/membresias.html
@@ -84,7 +84,7 @@
                     
                     <input type="text" class="inputText" id="nombre" name="nombre" pattern="[a-z]{2,20}" minlength="2" maxlength="20" placeholder="Ingresa el nombre" required>
                     <input type="text" class="inputText" id="duracion" name="duracion" pattern="[0-9]{1,3}" minlength="1" maxlength="3" placeholder="Ingresa la duracion en dias" required>
-                    <input type="text" class="inputText" id="precio" name="precio" pattern="[0-9]" placeholder="Ingresa el precio" required>
+                    <input type="number" class="inputText" id="precio" name="precio" pattern="[0-9]" minlength="1" placeholder="Ingresa el precio" required>
                     
                 </div>
                 <button type="submit" class="buttonSumit">Agregar</button>
